% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFPdataqualitycheck-package.R
\docType{package}
\name{WFPdataqualitycheck-package}
\alias{WFPdataqualitycheck}
\alias{WFPdataqualitycheck-package}
\title{WFPdataqualitycheck: Data Quality Check for World Food Programme Food Security Surveys}
\description{
A Shiny application for checking data quality and visualizing key food security indicators in World Food Programme (WFP) surveys. The app allows users to upload SPSS datasets and analyze Food Consumption Score (FCS), Household Dietary Diversity Score (HDDS), Reduced Coping Strategy Index (rCSI), Household Hunger Scale (HHS), and Livelihood Coping Strategy Index (LCS). Results can be viewed by different administrative levels and by enumerator.
}
\author{
\strong{Maintainer}: Idrissa Dabo \email{idrissa.dabo@wfp.org}

}
\keyword{internal}
